Split Checker.cs
