+++ Foxtrot/Foxtrot/Foxtrot.Extractor.csproj
@@ -149,6 +149,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Checker\Error.cs" />
+    <Compile Include="Checker\InspectorIncludingClosures.cs" />
+    <Compile Include="Checker\VisibilityHelper.cs" />
     <None Include="..\..\Microsoft.Research\common\Contracts.snk">
       <Link>Contracts.snk</Link>
     </None>
@@ -156,7 +159,7 @@
       <Link>version.cs</Link>
     </Compile>
     <Compile Include="AssemblyResolver.cs" />
-    <Compile Include="Checker.cs" />
+    <Compile Include="Checker\Warning.cs" />
     <Compile Include="CodeInspector.cs" />
     <Compile Include="ContractNodes.cs" />
     <Compile Include="Documentation.cs" />
@@ -164,6 +167,7 @@
     <Compile Include="Extractor.cs" />
     <Compile Include="ContractBlocks.cs" />
     <Compile Include="FindClosurePartsToDuplicate.cs" />
+    <Compile Include="Checker\PostExtractorChecker.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="HelperMethods.cs" />
   </ItemGroup>
